Redid Climb to be stabalized in code
Climb is now a PIDCommand, running a PID loop to maintain the roll of
the robot at the angle it is when Climb initializes. To provide access
to both roll and pitch, 2 methods, getRoll and getPitch were added to
Chassis. The parameter names for climb in Climber were modified to make
the two sides more clear, changing w1 and w2 to wL and wR respectively.